Use deps.cljs by default for NPM deps but fallback to project.clj
To maintain backwards compatibility so there are no breaking changes.

If a deps.cljs file exists in the project :source-paths then it will be
used for :npm-deps and :npm-dev-deps and anything in project.clj for the
same will be ignored.

If a deps.cljs file does not exist in the project :source-paths then
project.clj will be used for :npm-deps and :npm-dev-deps.